refactor(logic): split naval_resolution into part-of concern files (Refs #3290)
Phase-0 prerequisite slice for the colonizethis_logic package split epic.
Decomposes the 553-line lib/src/world/naval_resolution.dart (one of the
>500-line files Phase 0 calls out) into cohesive `part of` files, each well
under the 500-line target, with zero behaviour change.

Uses the same library/part-of pattern as the connectivity_resolver split on
this PR: identical public API and library-private symbols (incl. the shared
_NavalMoveOutcome typedef and _fleetIndexById helper); companion part files
inherit the library file's imports/exports.

naval_resolution.dart (553) -> four files:
- naval_resolution.dart (219): imports/exports, _NavalMoveOutcome, public
  entry points applyNavalMovesAndShipReveal + runNavalInterceptionCombatPhase
- naval_resolution_move.dart (251): dock/at-sea move handlers + reachability
- naval_resolution_battle.dart (68): victor resolution + dossier/dialogue
- naval_resolution_helpers.dart (44): fleet/sea-zone index + retreat helpers

Behaviour-preserving structural refactor (no SPEC change). Verified:
dart analyze (no issues on the four files), all 12 naval regression tests
pass, tool/check_logic_dead_files.dart recognizes the new part files.

Refs #3290
