+++ packages/colonizethis_logic/lib/src/world/naval_resolution.dart
@@ -25,6 +25,17 @@ export 'naval_coastal_visibility.dart'
         revealTilesAfterMoveToSeaZone;
 export 'naval_mission_orders.dart' show applyNavalMissionOrders;
 
+// Naval resolution concern fragments (Refs #3290 Phase-0 file-split). Each
+// `part of` fragment shares this library's imports and library-private scope,
+// so the move is behaviour-preserving — symbols, visibility, and helper
+// sharing (e.g. [_NavalMoveOutcome], [_fleetIndexById]) are unchanged. Splits a
+// 553-line file (one of the >500-line files Phase-0 calls out) into cohesive
+// files each well under the 500-line target ahead of the `colonizethis_world`
+// leaf-package extraction (Phase 1).
+part 'naval_resolution_helpers.dart';
+part 'naval_resolution_move.dart';
+part 'naval_resolution_battle.dart';
+
 /// Outcome of a single naval-move order application. Returned by both the
 /// dock and at-sea move handlers and consumed by [applyNavalMovesAndShipReveal]
 /// to thread the mutating fleet/visibility state across each player's orders.
@@ -35,289 +46,6 @@ typedef _NavalMoveOutcome = ({
   Map<String, int> fleetIndexById,
   Map<String, Map<String, String>> visibilityByTile,
 });
-
-/// Single-pass index: sea zone id → fleets whose [Fleet.seaZoneId] equals that
-/// zone. Preserves world fleet list order within each bucket (Refs #2394).
-Map<String, List<Fleet>> _fleetsBySeaZoneId(List<Fleet> fleets) {
-  final out = <String, List<Fleet>>{};
-  for (final f in fleets) {
-    final z = f.seaZoneId;
-    if (z == null) continue;
-    out.putIfAbsent(z, () => <Fleet>[]).add(f);
-  }
-  return out;
-}
-
-String? _firstFriendlyOrNeutralRetreatZone(
-  MapTopology topology,
-  String fromSeaZoneId,
-  String ownerId,
-  Map<String, Set<String>> hostileByOwner,
-  Map<String, List<Fleet>> fleetsBySeaZoneId,
-) {
-  for (final adj in nodesAdjacentTo(topology, fromSeaZoneId)) {
-    var hostileOwnersPresent = false;
-    for (final fleet in fleetsBySeaZoneId[adj] ?? const <Fleet>[]) {
-      if (!fleet.isAtSea) continue;
-      if (fleet.ownerId == ownerId) continue;
-      if (hostileByOwner[ownerId]?.contains(fleet.ownerId) ?? false) {
-        hostileOwnersPresent = true;
-        break;
-      }
-    }
-    if (!hostileOwnersPresent) return adj;
-  }
-  return null;
-}
-
-Map<String, int> _fleetIndexById(List<Fleet> fleets) => {
-  for (var i = 0; i < fleets.length; i++) fleets[i].id: i,
-};
-
-bool _navalMoveDestinationIsReachable({
-  required MapTopology topology,
-  required Fleet fleet,
-  required String destZoneId,
-}) {
-  if (fleet.isAtSea) {
-    final cur = fleet.seaZoneId;
-    if (cur == null) return false;
-    if (cur == destZoneId) return true;
-    return isAdjacentSeaSeaZone(topology, cur, destZoneId);
-  }
-  final inPortProvinceId = fleet.inPortAtProvinceId;
-  if (inPortProvinceId == null) return false;
-  final rl = regionAndLocalProvinceForFleetInPort(
-    inPortProvinceId,
-    fleet.regionId,
-  );
-  final provinceNodeId = provinceTopologyNodeId(
-    topology,
-    rl.localId,
-    rl.regionId,
-  );
-  if (provinceNodeId == null) return false;
-  return seaZonesAdjacentToProvince(
-    topology,
-    provinceNodeId,
-  ).contains(destZoneId);
-}
-
-_NavalMoveOutcome _applyDockNavalMoveOrder({
-  required Game game,
-  required MapTopology topology,
-  required List<Fleet> fleets,
-  required Map<String, Fleet> fleetById,
-  required Map<String, int> fleetIndexById,
-  required String playerId,
-  required String homeFleetId,
-  required Fleet fleet,
-  required NavalMoveOrder order,
-  required Map<String, Map<String, String>> visibilityByTile,
-}) {
-  final portProvinceId = order.destinationPortProvinceId!;
-  if (!fleet.isAtSea || fleet.seaZoneId == null) {
-    return (
-      fleets: fleets,
-      fleetById: fleetById,
-      fleetIndexById: fleetIndexById,
-      visibilityByTile: visibilityByTile,
-    );
-  }
-  final fullProvinceId = toFullProvinceId(fleet.regionId, portProvinceId);
-  final province = game.worldState.tryGetProvince(fullProvinceId);
-  if (province == null || province.ownerId != playerId) {
-    return (
-      fleets: fleets,
-      fleetById: fleetById,
-      fleetIndexById: fleetIndexById,
-      visibilityByTile: visibilityByTile,
-    );
-  }
-  final adjacentSeaZones = seaZoneIdsAdjacentToProvince(
-    topology,
-    fullProvinceId,
-  );
-  if (!adjacentSeaZones.contains(fleet.seaZoneId)) {
-    return (
-      fleets: fleets,
-      fleetById: fleetById,
-      fleetIndexById: fleetIndexById,
-      visibilityByTile: visibilityByTile,
-    );
-  }
-
-  var nextVis = revealProvinceTilesForPlayer(
-    game,
-    visibilityByTile,
-    playerId,
-    fullProvinceId,
-  );
-
-  if (dockOrderTargetsPlayerCapital(game, playerId, fullProvinceId)) {
-    final homeFleet = fleetById[homeFleetId];
-    if (homeFleet == null) {
-      return (
-        fleets: fleets,
-        fleetById: fleetById,
-        fleetIndexById: fleetIndexById,
-        visibilityByTile: nextVis,
-      );
-    }
-    final updatedHome = Fleet(
-      id: homeFleet.id,
-      ownerId: homeFleet.ownerId,
-      seaZoneId: null,
-      inPortAtProvinceId: homeFleet.inPortAtProvinceId,
-      regionId: homeFleet.regionId,
-      ships: [...homeFleet.ships, ...fleet.ships],
-      mission: FleetMission.none,
-      targetPortId: null,
-      targetProvinceId: null,
-    );
-    // Single-pass list build (Refs #2394): avoids intermediate lazy chains from
-    // `.where` / `.map` when merging a docking fleet into the capital home fleet.
-    final nextFleets = <Fleet>[
-      for (final f in fleets)
-        if (f.id != fleet.id) f.id == homeFleetId ? updatedHome : f,
-    ];
-    final nextFleetIndexById = _fleetIndexById(nextFleets);
-    fleetById[homeFleetId] = updatedHome;
-    fleetById.remove(fleet.id);
-    return (
-      fleets: nextFleets,
-      fleetById: fleetById,
-      fleetIndexById: nextFleetIndexById,
-      visibilityByTile: nextVis,
-    );
-  }
-
-  final portRegionId = ProvinceId.regionIdFrom(fullProvinceId);
-  final newFleet = Fleet(
-    id: fleet.id,
-    ownerId: fleet.ownerId,
-    seaZoneId: null,
-    inPortAtProvinceId: fullProvinceId,
-    regionId: portRegionId,
-    ships: fleet.ships,
-    mission: FleetMission.none,
-    targetPortId: null,
-    targetProvinceId: null,
-  );
-  var replacedDockFleet = false;
-  final nextFleets = <Fleet>[];
-  for (final f in fleets) {
-    if (f.id == fleet.id) {
-      nextFleets.add(newFleet);
-      replacedDockFleet = true;
-    } else {
-      nextFleets.add(f);
-    }
-  }
-  if (replacedDockFleet) {
-    fleetById[fleet.id] = newFleet;
-    return (
-      fleets: nextFleets,
-      fleetById: fleetById,
-      fleetIndexById: fleetIndexById,
-      visibilityByTile: nextVis,
-    );
-  }
-  return (
-    fleets: fleets,
-    fleetById: fleetById,
-    fleetIndexById: fleetIndexById,
-    visibilityByTile: nextVis,
-  );
-}
-
-_NavalMoveOutcome _applySeaNavalMoveOrder({
-  required Game game,
-  required MapTopology topology,
-  required List<Fleet> fleets,
-  required Map<String, Fleet> fleetById,
-  required Map<String, int> fleetIndexById,
-  required String playerId,
-  required Fleet fleet,
-  required NavalMoveOrder order,
-  required Map<String, Map<String, String>> visibilityByTile,
-}) {
-  final destZoneId = order.destinationSeaZoneId;
-  if (destZoneId == null || destZoneId.isEmpty) {
-    return (
-      fleets: fleets,
-      fleetById: fleetById,
-      fleetIndexById: fleetIndexById,
-      visibilityByTile: visibilityByTile,
-    );
-  }
-  if (!seaZoneNodeIds(topology).contains(destZoneId)) {
-    return (
-      fleets: fleets,
-      fleetById: fleetById,
-      fleetIndexById: fleetIndexById,
-      visibilityByTile: visibilityByTile,
-    );
-  }
-  if (!_navalMoveDestinationIsReachable(
-    topology: topology,
-    fleet: fleet,
-    destZoneId: destZoneId,
-  )) {
-    return (
-      fleets: fleets,
-      fleetById: fleetById,
-      fleetIndexById: fleetIndexById,
-      visibilityByTile: visibilityByTile,
-    );
-  }
-
-  final destRegionId = regionIdForSeaZone(topology, destZoneId);
-  final newFleet = Fleet(
-    id: fleet.id,
-    ownerId: fleet.ownerId,
-    seaZoneId: destZoneId,
-    inPortAtProvinceId: null,
-    regionId: destRegionId ?? fleet.regionId,
-    ships: fleet.ships,
-    mission: FleetMission.none,
-    targetPortId: null,
-    targetProvinceId: null,
-  );
-  var nextFleets = fleets;
-  var replacedAtSea = false;
-  final rebuiltAtSea = <Fleet>[];
-  for (final f in fleets) {
-    if (f.id == fleet.id) {
-      rebuiltAtSea.add(newFleet);
-      replacedAtSea = true;
-    } else {
-      rebuiltAtSea.add(f);
-    }
-  }
-  if (replacedAtSea) {
-    nextFleets = rebuiltAtSea;
-    fleetById[fleet.id] = newFleet;
-  }
-
-  var nextVis = visibilityByTile;
-  if (destRegionId != null) {
-    nextVis = revealTilesAfterMoveToSeaZone(
-      game: game,
-      topology: topology,
-      visibilityByTile: nextVis,
-      playerId: playerId,
-      destRegionId: destRegionId,
-      destZoneId: destZoneId,
-    );
-  }
-  return (
-    fleets: nextFleets,
-    fleetById: fleetById,
-    fleetIndexById: fleetIndexById,
-    visibilityByTile: nextVis,
-  );
-}
 
 Game applyNavalMovesAndShipReveal(
   Game game,
@@ -382,68 +110,6 @@ Game applyNavalMovesAndShipReveal(
       playerVisibilityByTile: visibilityByTile,
     ),
   );
-}
-
-String? _navalBattleWinnerOwnerId(
-  NavalBattleOutcome outcome,
-  BattleContextSea battle,
-) {
-  switch (outcome) {
-    case NavalBattleOutcome.side1Victory:
-      return battle.side1.ownerId;
-    case NavalBattleOutcome.side2Victory:
-      return battle.side2.ownerId;
-    case NavalBattleOutcome.stalemate:
-    case NavalBattleOutcome.mutualDestruction:
-      return null;
-  }
-}
-
-Game _applyNavalBattleVictoryDossierAndDialogue({
-  required Game state,
-  required BattleContextSea battle,
-  required NavalBattleResult result,
-  required int turn,
-  required int battleIndex,
-  required int seedAfterBattle,
-  void Function(DialogueEvent)? onDialogue,
-}) {
-  String? victorId;
-  String? loserId;
-  if (result.survivingShipsSide1.isEmpty &&
-      result.survivingShipsSide2.isNotEmpty) {
-    victorId = battle.side2.ownerId;
-    loserId = battle.side1.ownerId;
-  } else if (result.survivingShipsSide2.isEmpty &&
-      result.survivingShipsSide1.isNotEmpty) {
-    victorId = battle.side1.ownerId;
-    loserId = battle.side2.ownerId;
-  }
-  if (victorId == null || loserId == null) return state;
-
-  var next = state;
-  final evidence = evidenceForNavalBattleVictory(next, victorId, loserId, turn);
-  if (evidence.isNotEmpty) {
-    next = next.copyWith(
-      dossierEvidenceEntries: [...next.dossierEvidenceEntries, ...evidence],
-    );
-  }
-  final dialogueSeed =
-      (seedAfterBattle ^ (battleIndex * kTurnResolutionSeedMix)) &
-      kTurnResolutionLcgMask;
-  final events = dialogueEventsForNavalBattleResult(
-    next,
-    victorId,
-    loserId,
-    turn,
-    dialogueSeed,
-  );
-  if (onDialogue != null && events.isNotEmpty) {
-    for (final e in events) {
-      onDialogue(e);
-    }
-  }
-  return next;
 }
 
 Game runNavalInterceptionCombatPhase(
